Implemented first 3 powerups + some bombcollision

+++ Bowlerman/player.h
@@ -37,7 +37,7 @@ int getPlayerWidth();
 int getPlayerSpeed(Player p);
 int getPlayerID(Player p);
 void getStartPos(Player *p);
-void playerAddSpeed(Player *p, int speed);
+void playerAddSpeed(Player *p, float speed);
 void playerAddExplosionPower(Player *p, int explosionPower);
 void playerAddAmountOfBombs(Player *p, int amountOfBombs);
 PUBLIC void getStartPos(Player *p);
